hf_space/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
yfinance
torch
chronos-forecasting
